Change zettel anchor getting function definition
Unify the definitions of the two function to take a set of files instead
of different arguments. The list of files has the form:
files = { "full/path/name.md" = "name.md" }

I.e. the complete path is the key and the file basename the value,
leading to a set of files, even if basenames are the same.

Usually it will be created by the get_all_files function of the same
module, but it can be passed in manually as well (e.g. to avoid
duplicate filesystem lookups, etc).